Rename package dbc to can_dbc_parser for a more descriptive name
