+++ pacifica_dbw/package.xml
@@ -15,13 +15,13 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
+  <exec_depend>can_dbc_parser</exec_depend>
   <exec_depend>dbw_pacifica_can</exec_depend>
   <exec_depend>dbw_pacifica_joystick_demo</exec_depend>
   <exec_depend>dbw_pacifica_joystick_speed_demo</exec_depend>
   <exec_depend>dbw_pacifica_msgs</exec_depend>
   <exec_depend>pdu</exec_depend>
   <exec_depend>pdu_msgs</exec_depend>
-  <exec_depend>dbc</exec_depend>
 
   <export>
     <metapackage/>
