+++ pdu/package.xml
@@ -20,7 +20,7 @@
   <depend>std_msgs</depend>
   <depend>can_msgs</depend>
   <depend>pdu_msgs</depend>
-  <depend>dbc</depend>
+  <depend>can_dbc_parser</depend>
 
   <export>
   </export>
