+++ pdu/src/pdu.h
@@ -46,10 +46,10 @@
 #include <std_msgs/Bool.h>
 #include <std_msgs/String.h>
 
-#include <dbc/DbcMessage.h>
-#include <dbc/DbcSignal.h>
-#include <dbc/Dbc.h>
-#include <dbc/DbcBuilder.h>
+#include <can_dbc_parser/DbcMessage.h>
+#include <can_dbc_parser/DbcSignal.h>
+#include <can_dbc_parser/Dbc.h>
+#include <can_dbc_parser/DbcBuilder.h>
 
 namespace NewEagle
 {
